Simplify code for adding and looking up mixins.
Fix bug where adding a double sided mixin followed by a single sided mixin (with the same key) caused the second mixin to be double sided. Eg/

"to": "TO", // Mixin that works on both left and right sides ("double sided")
"to|mr": "-TD" // Mixin that only works on right side ("single sided")

caused "-TD" to work on both the left and right side instead of only the right side.